Update README.md
# Hit the Target !
It is a simple game where you have to target a dart to the active board . The speed of the game increases with time. You can choose difficulty levels which changes the number of dart board.  Enjoy playing :)

## Technologies used
1. React.js
2. JavaScript
3. HTML
4. CSS

## Setup and usage

Clone repository
Install dependencies with npm install
npm start to run locally

## Live page: https://speedgame-at.netlify.app/

## Authors and acknowledgment

Anupa.